[FEAT] Diet box 컴포넌트화 (#80)
